// routes/logic/classes/interfaces/StoragePlacesInterfaces.ts
export interface StoragePlacesInterface {
  storagePlacesId: number;
  storagePlacesLoc: string;
  storagePlacesWhloc: string;
  storagePlacesStatusPos: string;
  storagePlacesVirtualLoc: number;
}

export interface StoragePlacesDataInterface {
  ID:number;
  LOC: string;
  WHLOC: string;
  STATUS_POS: string;
  VIRTUAL_LOC: number;
}